extract OCD source class
Split Source::OCD out into its own file
