+++ lib/source/ocd.rb
@@ -1,0 +1,23 @@
+require_relative 'csv'
+
+module Source
+  class OCD < CSV
+    def fields
+      %i(area area_id)
+    end
+
+    def fuzzy_match?
+      i(:merge)[:fuzzy]
+    end
+
+    def overrides
+      return {} unless i(:merge)
+      return {} unless i(:merge).key? :overrides
+      i(:merge)[:overrides]
+    end
+
+    def generate
+      i(:generate)
+    end
+  end
+end
